Update docs.md
Yesterday on the chat I discussed with @adi8i and @w00fz that I should add a lot of things to the docs. Everything that I can add was welcome. So hereby I make my very first GitHub fork/change. Hopefully I made no typo's and it gets approved so that in future I can add a lot more. I need a bit of help with the markup but I am sure that this will be quickly learned. Basically I want the developers to develop and let us (community) take care of the docs. We can do that.

SO WHAT IS CHANGED:
Added two examples under HTML Input Field Types
(I wanted to do all but lets see first if this is oke)
Added a new section called: Gantry Alerts, Notes and Styling. In this section I explained the use of the seperator.note which w00fz wanted to see in the docs and I promised to put it in.