Create 09.using-gantry-syntax.md
As I am receiving a lot of question about using Gantry Syntax in Custom HTML modules I went on and created a section for it in the docs. As nr. 9 was missing I created it under that number. Please review the location maybe this fits better somewhere else. I though this would be the best fit. I will be adding more in future but would like to see this get merged with the main branch a.s.a.p that way I don't keep repeating myself on the chat.