There is no such class as g-row is there?  We use g-grid
You have called it g-row which sounds like a cross between Gantry & Bootstrap.  I believe this is incorrect but please tell me off if I'm wrong.  It probably was this at one stage, but now we use g-grid, and all of the example html only shows g-grid not g-row.  If g-row can be used can you please include a HTML example on this page.