Fixed html code for example 2
The first `<div>` had a link at the end of the text that wasn't closed
with a `>`. Further inspection showed it was an empty link with a `.` at
the end which also followed a `?`.

I added the missing `>`, placed the text inside the link, and removed
the unnecessary period `.` at the end.