A lot of people ask about section styles in the forum
I've seen quite a few questions about section styles in the forum.  Although usually the best advice is to use CSS (and I have included this advice) I still wanted to know the full way of doing things and I think some others would like to know this.  To save myself searching the forums and to save you guys answering the same questions (and support is excellent) I decided to improve this help document.  I personally feel that this would be better on it's own page as it is not a preset but I don't have the ability to do that.  I hope that if you decide to take the route I have at least saved you some time.